// casetudy/src/app/models/service/service.ts
export interface Service {
  id?: number;
  name?: string;
  image?: string;

}

// let service1: Service = {
//   name: "Phòng hoa hồng",
//   image:"https://furamavietnam.com/wp-content/uploads/2018/03/Vietnam_Danang_Furama_Ocean-Suite-Feature-370x239.jpg",
//
// }
// let service2:Service = {
//   name: "Phòng hoa hồng",
//   image:"https://furamavietnam.com/wp-content/uploads/2018/03/Vietnam_Danang_Furama_Ocean-Suite-Feature-370x239.jpg",
//
// }
// let service3:Service = {
//   name: "Phòng hoa hồng",
//   image:"https://furamavietnam.com/wp-content/uploads/2018/03/Vietnam_Danang_Furama_Ocean-Suite-Feature-370x239.jpg",
//
// }
// let service4:Service = {
//   name: "Phòng hoa hồng",
//   image:"https://furamavietnam.com/wp-content/uploads/2018/03/Vietnam_Danang_Furama_Ocean-Suite-Feature-370x239.jpg",
//
// }
// let service5:Service = {
//   name: "Phòng hoa hồng",
//   image:"https://furamavietnam.com/wp-content/uploads/2018/03/Vietnam_Danang_Furama_Ocean-Suite-Feature-370x239.jpg",
//
// }
// let service6:Service = {
//   name: "Phòng hoa hồng",
//   image:"https://furamavietnam.com/wp-content/uploads/2018/03/Vietnam_Danang_Furama_Ocean-Suite-Feature-370x239.jpg",
//
// }
// let array :Service[]=[]
// array.push(service1);
// array.push(service2);
// array.push(service3);
// array.push(service4);
// array.push(service5);
// array.push(service6);